Introduce sharedParameters between windows
This is an effort to enable parameter update in child windows.
For example, suppose the "on-pulse IDs" is modified in the main GUI,
once the "run" button is clicked, all the child windows should
start to use the updated "on-pulse IDs". This is implemented in the
abstract class so there is no need to take care of them for individual
windows. There are still problems to be addressed:

1. Find a solution to update ParameterTree automatically;
2. Provide a "freeze" option for the window, i.e. the shared parameters
   in a child window can block any update, since we may want to use
   different shared paramters in one run.