Handle Threads and Processes properly at anytime
Implement a TimeoutClient by subclassing Client in karabo_bridge in
order to have a timeout option for client. Both DaqWorker and
DataProcessor will be instantiated in __init__ method of MainGUI only
once. Moreover, the two threads will be cleaned no matter the GUI is
terminated normally or unnormally. In addition, the FileServer is also
taken care of properly in both situations.

Other modifications in this commit:

- Automatically register plot windows/widgets.

- Remove abstract method updatePlots and clearPlots from
abstract class AbstractWindow.