Display difference curve in another window.
In LaserOnOffWindow, the difference curve (On - off) is displayed in
it own window since this curve is most interested in some experiment
and it should stay there if no valid data is received.

Also rename fom_range to diff_integration_range. "FOM range" will be
displayed as "Diff integration range" to avoid confusion.