Replace method poni with pos_inv and rename PONI in the code
My understanding about PONI is wrong. In the pyFAI instruction, the
origin of PONI is the lower-left corner of the image. However, the
image defines the (0, 0) point at the lower-left corner. In usual
image software, we have (0, 0) point at the upper-left corner, and
pyFAI uses exactly the upper-left corner as the origin!!!