Add initial support for saving and restoring sessions
This has been requested because sometimes extra-foam will crash (particularly on
long-running instances) and the users will lose all settings they've set. These
can be very tedious and error-prone to redo, hence the desire to restore the
settings from the previous session.

Doing this for _all_ settings in extra-foam is a huge amount of work, so this
initial implementation only supports restoring the ROIs, which is one of the
more important settings.