Update pyzmq to fix a memory leak
A memory leak (or something that at least behaves like one) was observed
whenever streaming AGIPD data to extra-foam. This only seemed to happen with
AGIPD (also tested with EPIX/DSSC) and only on developer machines, never on the
online cluster. Memory usage would steadily increase in the pulse worker
process, until it would be killed by the OOM killer after a couple of
minutes. Using gc.collect() in the pulse worker did fix this, but significantly
slowed things down.

I initially thought this was something to do with one of the processors in the
pulse worker, but disabling all those did nothing. Eventually I tracked it down
to the karabo_bridge.Client which reads data from the streaming sources. Note
that EXtra-foam has a somewhat advanced proxy (BridgeProxy) that can connect to
multiple backends (Karabo bridges), and place them all into a single queue for
the Client. Directly connecting the Client to a single backend did nothing, and
I was able to reproduce the memory leak with a very simple standalone script
using the Client to read data from a bridge and immediately discard it.

Turns out that it was fixed by upgrading pyzmq. We appear to have run into this
bug: https://github.com/zeromq/pyzmq/issues/1582
TL;DR: there are multiple levels of garbage collection, and some pyzmq objects
were only cleaned up on a full (level 2) GC, which doesn't happen frequently
enough in some circumstances.

What I believe was happening is that the large size of the AGIPD
data caused memory usage to increase rapidly, and on the online cluster nodes
this is ok because they have tons of memory and can afford to wait around for
the full GC, but our developer machines do not have that much and simply run out
before the full GC occurs. Anecdotally, while debugging this with
`gc.get_stats()` I only ever recall seeing 0th and 1st level GC runs, never 2nd
level.

Luckily the bug was fixed in version 22.3.0, which was released today:
- https://github.com/zeromq/pyzmq/pull/1590