Fix handling of raw multi-module detector data during train matching
Multi-module detectors are handled specially during train matching because they
have multiple source entries in the data received from a Karabo bridge. The code
was creating entries for the internal raw data cache without first checking that
that the raw data existed in the data from the bridge. This caused train
matching to break if data from a detector is not present on every train.