Add support for special suites supporting both EXtra-foam and Karabo
Some special suites may want to support processing data from both EXtra-foam and
Karabo bridges, this is now possible by passing ClientType.BOTH to the
constructor of _SpecialAnalysisBase(). More generally, this replaces the
`client_klass` parameter of create_special() with an argument to the
_SpecialAnalysisBase() constructor.

If both client types are supported, then a combo box will be available for the
user to select the appropriate source.