Convert floats to ints for Python 3.10 compatibility
Python 3.10 included this change: https://bugs.python.org/issue37999. Turns out
that extra-foam's copy of pyqtgraph relies on this in some places, which results
in TypeError's when running it with Python 3.10. These are just the places in
the code that I could find with manual/unit tests, it's possible that there are
more. The plan is to replace the internal copy with upstream though, so I don't
think it's worth doing anything more rigorous.